% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tree.R
\name{draw_tree}
\alias{draw_tree}
\title{Choose cells interactively to subset a cds}
\usage{
draw_tree(
  cds,
  reduction_method = c("UMAP", "tSNE", "PCA", "Aligned"),
  color_cells_by = "cluster"
)
}
\arguments{
\item{input}{CDS object to subset}
}
\value{
A subset CDS object. If return_list = FALSE, a list of cell names.
}
\description{
Choose cells interactively to subset a cds
}
